refactor: InfoLayout -> ProfileLayout
